Get to a nicely broken spot with live-reload server
This is a great example of code that is broken without some kind of
explicitly coordination: shutdown does not work! You can hit `^C` as
many times as you want and it will never do anything, because the axum
server does not get the message (…literally: there is no message for it
to get, as this stands).