lrrs: tweak channel design a bit
I am not *totally* happy with this, but it eliminates one of the major
ways things were janky before: the watcher sends a change specifically
and only about file changes, *not* one of the generic messages. That
means it does not have to be handled elsewhere, and the one `select!`
which refers to it can be much more narrow as well.

The reason I am not in love with it is: it raelly feels like it would be
nice to just send over a normal `mpsc::Receiver<FileChanged>` instead.
Unfortunately, that hits ownership problems unless the state design is
significantly different, because with it sitting behind an `Arc`, it is
not available for mutable access, which an `mpsc::Receiver` requires to
do the `.recv()` call.