lrrs: handle SendError for the FileChanged Sender
as things are configured at present, if there is no `WebSocket`
configured yet, the other end of this fails to do anything, because
there is no receiver. Per the docs:

> A send operation can only fail if there are no active receivers,
> implying that the message could never be received. The error contains
> the message being sent as a payload so it can be recovered.

What this means is: I need (a) to handle that error, but also (b) to
think about the relationships between the parts of the system. There are
two basic ways to approach this, after all:

1. Treat a `SendError` as a case to possibly log some info (e.g. with
   `tracing` or such), but an otherwise perfectly normal case which can
   safely be ignored.
2. Treat it as a smell in the design. Why is it even *trying* to send that
   message if there is nothing listening?

Here, opt for (2): if there are no `Receiver`s (either from `.subscribe`
calls or from the original channel creation), do not send a message
across the channel at all, even if there is a change.

(As an aside, it would be really nice if  you *could* guarantee that
there were something on the other end, but since it is dynamic, you
really cannot translate it into the types of the program, I think.)