lrrs: get rid of *alllll* the extra cruft
This was way, way, *way* overcomplicated before. Since the server does
not care about messages from the `WebSocket` instance other than making
sure not to send messages when it is closed, the internal messages were
all irrelevant. This really just needs to let the watcher send messages
to any and all subscribed `WebSocket`s. So… now that is all it does!