lr2: duh, just need to not drop the watcher!
In my previous commit, I was being silly and trying to think of how to
interact in some long-lived way with the watcher, to avoid dropping it.
Here, I do the obvious thing: return it, and don't drop it until the
various async operations are all done. That’s it! That’s the whole
“solution” to the “problem” I was having last time.

The key insight for me was that it needs to live on, and that the way
e.g. `watchexec` does it makes that be true by way of a `loop` because
(and really only because) it is watching for changes to the *config*,
and that loop went on indefinitely `.await`-ing changes to that config
so that the watcher could be updated. This small and simple example
neither has nor needs that complexity.