lr2: further (and fairly dramatically!) simplify watcher
I was, again, just plain overthinking this. There is no need to spawn a
task here at all: it is already in an `async fn`, so it will be run in a
task automatically by its *parent*. Additionally, pulling the call to
`watcher.watch()` out into the `watch` makes this much easier to follow.