Example of retry and timeout implementations
I have one bit of wrong type juggling (Option instead of Result) here
but the overall shape of this is what I want and I can finish it up at a
future point; it mostly gives me a sense of whether it is reasonable to
implement these in terms of the underlying futures primitives as part of
the chapter (maybe in the final section). Maybe?

../notes/Book/Runtime decision.md JJ:     A Cargo.lock JJ:     A
Cargo.toml JJ:     A src/main.rs